chore(kc admin cl): fix split packages and refactor reactive names
