--- extensions/keycloak-admin-client-common/deployment/src/main/java/io/quarkus/keycloak/admin/client/common/KeycloakAdminClientInjectionEnabled.java
@@ -1,13 +1,0 @@
-package io.quarkus.keycloak.admin.client.common;
-
-import java.util.function.BooleanSupplier;
-
-public class KeycloakAdminClientInjectionEnabled implements BooleanSupplier {
-
-    KeycloakAdminClientBuildTimeConfig config;
-
-    @Override
-    public boolean getAsBoolean() {
-        return config.enabled();
-    }
-}
