+++ extensions/keycloak-admin-client-common/deployment/src/main/java/io/quarkus/keycloak/admin/client/common/deployment/KeycloakAdminClientInjectionEnabled.java
@@ -1,0 +1,13 @@
+package io.quarkus.keycloak.admin.client.common.deployment;
+
+import java.util.function.BooleanSupplier;
+
+public class KeycloakAdminClientInjectionEnabled implements BooleanSupplier {
+
+    KeycloakAdminClientBuildTimeConfig config;
+
+    @Override
+    public boolean getAsBoolean() {
+        return config.enabled();
+    }
+}
