--- extensions/keycloak-admin-rest-client/runtime/src/main/java/resources/META-INF/quarkus-extension.yaml
@@ -1,14 +1,0 @@
----
-artifact: ${project.groupId}:${project.artifactId}:${project.version}
-name: "Reactive Keycloak Admin Client"
-metadata:
-  keywords:
-  - "keycloak"
-  - "keycloak-admin-client"
-  - "admin"
-  - "openid-connect"
-  - "reactive"
-  categories:
-    - "security"
-    - "reactive"
-  status: "stable"
